third_party/blink/web_tests/platform/generic/external/wpt/speech-api/SpeechSynthesisUtterance-basics.https-expected.txt

This is a testharness.js-based test.
PASS new SpeechSynthesisUtterance() default text
PASS new SpeechSynthesisUtterance() default lang
PASS new SpeechSynthesisUtterance() default voice
FAIL new SpeechSynthesisUtterance() default volume assert_equals: volume expected 1 but got -1
FAIL new SpeechSynthesisUtterance() default rate assert_equals: rate expected 1 but got -1
FAIL new SpeechSynthesisUtterance() default pitch assert_equals: pitch expected 1 but got -1
FAIL new SpeechSynthesisUtterance("hello") text and defaults assert_equals: volume expected 1 but got -1
PASS new SpeechSynthesisUtterance(null)
PASS new SpeechSynthesisUtterance(undefined)
PASS SpeechSynthesisUtterance text setter
Harness: the test ran to completion.
